backend/requirements.txt
Django==3.2.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-decouple==3.4
djangorestframework==3.5
django-cors-headers==3.5